backend/requirements.txt
aiofiles==23.2.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
brotli==1.2.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cssselect2==0.8.0
fastapi==0.123.5
fonttools==4.61.0
greenlet==3.2.4
h11==0.16.0
html5lib==1.1
httptools==0.7.1
idna==3.11
Jinja2==3.1.2
joblib==1.5.2
MarkupSafe==3.0.3
nltk==3.8.1
numpy==2.3.5
pandas==2.3.3
pillow==12.0.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydyf==0.12.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
SQLAlchemy==2.0.44
starlette==0.50.0
threadpoolctl==3.6.0
tinycss2==1.5.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
watchfiles==1.1.1
weasyprint==60.2
webencodings==0.5.1
websockets==15.0.1
zopfli==0.4.0
